+++ tests/IntegrationTests/ApplicationTestData.cs
@@ -1,0 +1,155 @@
+// ------------------------------------------------------------
+//  Copyright (c) Microsoft Corporation.  All rights reserved.
+//  Licensed under the MIT License (MIT). See License.txt in the repo root for license information.
+// ------------------------------------------------------------
+
+
+using System.Text.RegularExpressions;
+using Microsoft.Azure.IIoT.Auth.Clients;
+using Microsoft.Azure.IIoT.Diagnostics;
+using Opc.Ua;
+using Opc.Ua.Gds;
+using Opc.Ua.Test;
+
+namespace Microsoft.Azure.IIoT.OpcUa.Services.Vault.Test
+{
+    public class ApplicationTestData
+    {
+        public ApplicationTestData()
+        {
+            Initialize();
+        }
+
+        private void Initialize()
+        {
+            ApplicationRecord = new ApplicationRecordDataType();
+            CertificateGroupId = null;
+            CertificateTypeId = null;
+            CertificateRequestId = null;
+            DomainNames = new StringCollection();
+            Subject = null;
+            PrivateKeyFormat = "PFX";
+            PrivateKeyPassword = "";
+            Certificate = null;
+            PrivateKey = null;
+            IssuerCertificates = null;
+        }
+
+        public ApplicationRecordDataType ApplicationRecord;
+        public NodeId CertificateGroupId;
+        public NodeId CertificateTypeId;
+        public NodeId CertificateRequestId;
+        public StringCollection DomainNames;
+        public string Subject;
+        public string PrivateKeyFormat;
+        public string PrivateKeyPassword;
+        public byte[] Certificate;
+        public byte[] PrivateKey;
+        public byte[][] IssuerCertificates;
+    }
+
+    public class ApplicationTestDataGenerator
+    {
+
+        public ApplicationTestDataGenerator(int randomStart = 1)
+        {
+            _randomStart = randomStart;
+            _randomSource = new RandomSource(_randomStart);
+            _dataGenerator = new DataGenerator(_randomSource);
+        }
+
+        public ApplicationTestData RandomApplicationTestData()
+        {
+            ApplicationType appType = (ApplicationType)_randomSource.NextInt32((int)ApplicationType.ClientAndServer);
+            string pureAppName = _dataGenerator.GetRandomString("en");
+            pureAppName = Regex.Replace(pureAppName, @"[^\w\d\s]", "");
+            string pureAppUri = Regex.Replace(pureAppName, @"[^\w\d]", "");
+            string appName = "UA " + pureAppName;
+            StringCollection domainNames = RandomDomainNames();
+            string localhost = domainNames[0];
+            string privateKeyFormat = _randomSource.NextInt32(1) == 0 ? "PEM" : "PFX";
+            string appUri = ("urn:localhost:opcfoundation.org:" + pureAppUri.ToLower()).Replace("localhost", localhost);
+            string prodUri = "http://opcfoundation.org/UA/" + pureAppUri;
+            StringCollection discoveryUrls = new StringCollection();
+            StringCollection serverCapabilities = new StringCollection();
+            switch (appType)
+            {
+                case ApplicationType.Client:
+                    appName += " Client";
+                    break;
+                case ApplicationType.ClientAndServer:
+                    appName += " Client and";
+                    goto case ApplicationType.Server;
+                case ApplicationType.Server:
+                    appName += " Server";
+                    int port = (_dataGenerator.GetRandomInt16() & 0x1fff) + 50000;
+                    discoveryUrls = RandomDiscoveryUrl(domainNames, port, pureAppUri);
+                    break;
+            }
+            ApplicationTestData testData = new ApplicationTestData
+            {
+                ApplicationRecord = new ApplicationRecordDataType
+                {
+                    ApplicationNames = new LocalizedTextCollection { new LocalizedText("en-us", appName) },
+                    ApplicationUri = appUri,
+                    ApplicationType = appType,
+                    ProductUri = prodUri,
+                    DiscoveryUrls = discoveryUrls,
+                    ServerCapabilities = serverCapabilities
+                },
+                DomainNames = domainNames,
+                Subject = string.Format("CN={0},DC={1},O=OPC Foundation", appName, localhost),
+                PrivateKeyFormat = privateKeyFormat
+            };
+            return testData;
+        }
+
+        private string RandomLocalHost()
+        {
+            string localhost = Regex.Replace(_dataGenerator.GetRandomSymbol("en").Trim().ToLower(), @"[^\w\d]", "");
+            if (localhost.Length >= 12)
+            {
+                localhost = localhost.Substring(0, 12);
+            }
+            return localhost;
+        }
+
+        private string[] RandomDomainNames()
+        {
+            int count = _randomSource.NextInt32(8) + 1;
+            string[] result = new string[count];
+            for (int i = 0; i < count; i++)
+            {
+                result[i] = RandomLocalHost();
+            }
+            return result;
+        }
+
+        private StringCollection RandomDiscoveryUrl(StringCollection domainNames, int port, string appUri)
+        {
+            StringCollection result = new StringCollection();
+            foreach (string name in domainNames)
+            {
+                int random = _randomSource.NextInt32(7);
+                if ((result.Count == 0) || (random & 1) == 0)
+                {
+                    result.Add(string.Format("opc.tcp://{0}:{1}/{2}", name, (port++).ToString(), appUri));
+                }
+                if ((random & 2) == 0)
+                {
+                    result.Add(string.Format("http://{0}:{1}/{2}", name, (port++).ToString(), appUri));
+                }
+                if ((random & 4) == 0)
+                {
+                    result.Add(string.Format("https://{0}:{1}/{2}", name, (port++).ToString(), appUri));
+                }
+            }
+            return result;
+        }
+
+        private int _randomStart = 1;
+        private RandomSource _randomSource;
+        private DataGenerator _dataGenerator;
+    }
+
+}
